Added a drawnow at the end of each script
The system would freeze for me because the figure buffers were not being flushed as we ran through some of the tutorials.  I added the drawnow to flush every one of the tutorials or validations.  That was recommended on the web, and it did work for me on my (slightly older) machine.